requirements.txt
docxtpl
PyYAML
